first pass at bug fix
